% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haunted_data.R
\docType{data}
\name{haunted_data}
\alias{haunted_data}
\title{Haunted Places in the U.S.}
\format{
A data frame with several 10,992 rows and 11 columns, each describing a unique haunted location. The variables are as follows:
\describe{
\item{city}{Name of the city where the haunted location is situated}
\item{location}{Name or description of the haunted location}
\item{state}{State abbreviation where the city is located}
\item{country}{Country where the location is found (e.g., "United States")}
\item{description}{Brief description of the haunting or history of the location}
\item{latitude}{Latitude of the haunted location}
\item{longitude}{Longitude of the haunted location}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2023/2023-10-10/haunted_places.csv'}
}
\usage{
haunted_data
}
\description{
A dataset that contains information about haunted locations in various cities
across the United States.
}
\keyword{datasets}
